@@ -1,8 +1,8 @@
 module github.com/fsgo/fspool
 
-go 1.19
+go 1.20
 
-require github.com/stretchr/testify v1.8.0
+require github.com/stretchr/testify v1.8.4
 
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
